Factored out release generation commands
